Fix missing status variable reference
